# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "BlogFlow API",
        "description": "API REST de la plateforme BlogFlow. Authentification via token Bearer (Laravel Sanctum). Obtenez votre token via POST /api/login.",
        "contact": {
            "name": "Équipe BlogFlow",
            "email": "admin@blogflow.fr"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://blog-api-service-fbnq.onrender.com",
            "description": "Serveur de production"
        },
        {
            "url": "http://localhost:8000",
            "description": "Serveur local"
        }
    ],
    "paths": {
        "/api/articles": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Liste paginée des articles",
                "description": "Retourne 10 articles par page avec leurs catégories et auteurs. Supporte la recherche plein texte et le filtrage par catégorie.",
                "operationId": "11aa4cc6561a7b94c56d1201d62513ee",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Numéro de page (défaut : 1)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Recherche dans le titre et le contenu",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "react"
                        }
                    },
                    {
                        "name": "category_id",
                        "in": "query",
                        "description": "Filtrer par identifiant de catégorie",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PaginatedArticles"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Articles"
                ],
                "summary": "Créer un article",
                "description": "Crée un nouvel article. L'auteur est automatiquement l'utilisateur authentifié.",
                "operationId": "e328bed23caf889ad982ad3f1180bccb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title",
                                    "short_description",
                                    "content",
                                    "category_id"
                                ],
                                "properties": {
                                    "title": {
                                        "description": "Max 191 caractères",
                                        "type": "string",
                                        "example": "Mon nouvel article"
                                    },
                                    "short_description": {
                                        "description": "Max 255 caractères",
                                        "type": "string",
                                        "example": "Description courte"
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "Contenu complet de l'article"
                                    },
                                    "category_id": {
                                        "description": "ID d'une catégorie existante",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "image": {
                                        "description": "jpeg/png/jpg/gif, max 2 Mo",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Article créé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Article créé avec succès !"
                                        },
                                        "article": {
                                            "$ref": "#/components/schemas/Article"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "User": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Sophie Marchand"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "admin@blogflow.fr"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Category": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Développement Web"
                    }
                },
                "type": "object"
            },
...
            "Article": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Les fondamentaux de React 19"
                    },
                    "short_description": {
                        "type": "string",
                        "example": "React 19 introduit des changements profonds..."
                    },
                    "content": {
                        "type": "string",
                        "example": "Contenu complet de l'article..."
                    },
                    "image": {
                        "type": "string",
                        "example": "https://blog-api-service-fbnq.onrender.com/storage/articles/photo.jpg",
                        "nullable": true
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "category_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "category": {
                        "$ref": "#/components/schemas/Category"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    }
                },
                "type": "object"
            },
...
            "PaginatedArticles": {
                "properties": {
                    "current_page": {
                        "type": "integer",
                        "example": 1
                    },
                    "last_page": {
                        "type": "integer",
                        "example": 2
                    },
                    "per_page": {
                        "type": "integer",
                        "example": 10
                    },
                    "total": {
                        "type": "integer",
                        "example": 12
                    },
                    "next_page_url": {
                        "type": "string",
                        "nullable": true
                    },
                    "prev_page_url": {
                        "type": "string",
                        "nullable": true
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Article"
                        }
                    }
                },
                "type": "object"
            },
            "ValidationError": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The title field is required."
                    },
                    "errors": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
}